Move content from document to slug
