+++ client/src/pages/_sites/[site]/[[...slug]].tsx
@@ -33,6 +33,14 @@ const API_ENDPOINT = process.env.API_ENDPOINT;
 if (typeof window !== 'undefined' && !('ResizeObserver' in window)) {
   window.ResizeObserver = ResizeObserver;
 }
+
+const FAVICON_VERSION = 3;
+
+function v(href: string) {
+  // To do, come back and set the basepath
+  return `${href}?v=${FAVICON_VERSION}`;
+}
+
 // TODO - Add ProgessBar back when you can access color. (Put inside Page component?)
 // const progress = new ProgressBar({
 //   size: 2,
@@ -96,6 +104,49 @@ export default function Page({ stringifiedMdxSource, stringifiedData }: PageProp
             <ColorVariables />
             <Title suffix={config.name}>{title}</Title>
             <Head>
+              <link
+                rel="apple-touch-icon"
+                sizes="180x180"
+                href={v('/favicons/apple-touch-icon.png')}
+              />
+              <link
+                rel="icon"
+                type="image/png"
+                sizes="32x32"
+                href={v('/favicons/favicon-32x32.png')}
+              />
+              <link
+                rel="icon"
+                type="image/png"
+                sizes="16x16"
+                href={v('/favicons/favicon-16x16.png')}
+              />
+              <link rel="shortcut icon" href={v('/favicons/favicon.ico')} />
+              <meta name="apple-mobile-web-app-title" content={config.name} />
+              <meta name="application-name" content={config.name} />
+              <meta name="theme-color" content="#ffffff" />
+              <meta name="msapplication-TileColor" content={config.colors?.primary} />
+              <meta name="msapplication-config" content={v('/favicons/browserconfig.xml')} />
+              <meta name="theme-color" content="#ffffff" />
+              <script
+                dangerouslySetInnerHTML={{
+                  __html: `
+                try {
+                  if (localStorage.theme === 'dark' || (${(
+                    config.modeToggle?.default == null
+                  ).toString()} && !('theme' in localStorage) && window.matchMedia('(prefers-color-scheme: dark)').matches) || ${(
+                    config.modeToggle?.default === 'dark'
+                  ).toString()}) {
+                    document.documentElement.classList.add('dark')
+                  }
+                  
+                  else {
+                    document.documentElement.classList.remove('dark')
+                  }
+                } catch (_) {}
+              `,
+                }}
+              />
               {config?.metadata &&
                 Object.entries(config?.metadata).map(([key, value]) => {
                   if (!value) {
@@ -109,21 +160,31 @@ export default function Page({ stringifiedMdxSource, stringifiedData }: PageProp
             </Head>
             <GA4Script ga4={analyticsConfig.ga4} />
             <SearchProvider>
-              <Header
-                hasNav={Boolean(config.navigation?.length)}
-                navIsOpen={navIsOpen}
-                onNavToggle={(isOpen: boolean) => setNavIsOpen(isOpen)}
-                title={meta?.title}
-                section={section}
-              />
-              <DocumentationLayout
-                nav={nav}
-                navIsOpen={navIsOpen}
-                setNavIsOpen={setNavIsOpen}
-                meta={meta}
+              <div
+                className="antialiased bg-background-light dark:bg-background-dark text-slate-500 dark:text-slate-400"
+                // Add background image
+                {...(config.backgroundImage && {
+                  style: {
+                    background: `url('${config.backgroundImage}') no-repeat fixed top right`,
+                  },
+                })}
               >
-                <MDXRemote components={components} {...mdxSource} />
-              </DocumentationLayout>
+                <Header
+                  hasNav={Boolean(config.navigation?.length)}
+                  navIsOpen={navIsOpen}
+                  onNavToggle={(isOpen: boolean) => setNavIsOpen(isOpen)}
+                  title={meta?.title}
+                  section={section}
+                />
+                <DocumentationLayout
+                  nav={nav}
+                  navIsOpen={navIsOpen}
+                  setNavIsOpen={setNavIsOpen}
+                  meta={meta}
+                >
+                  <MDXRemote components={components} {...mdxSource} />
+                </DocumentationLayout>
+              </div>
             </SearchProvider>
           </AnalyticsContext.Provider>
         </ConfigContext.Provider>
